Corrected a few issues, and rewritten the PIR code
